requirements.txt
# Core dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.3.7

# Vision and image processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# Text-to-speech
pyttsx3==2.90
langdetect==1.0.9

# HTTP requests
requests==2.31.0

# Hugging Face models and transformers
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
datasets>=2.14.0
tokenizers>=0.14.0

# Qwen model dependencies
qwen-tokenizer>=1.0.0
einops>=0.7.0

# Additional HF dependencies
sentencepiece>=0.1.99
protobuf>=3.20.0

# Optional: for better performance
# torchvision>=0.15.0  # Uncomment if using CUDA
# torchaudio>=2.0.0    # Uncomment if using CUDA 